RAG_script/requirements.txt
langchain>=0.3.0
langchain-openai>=0.3.0
langchain-community>=0.3.0
openai>=1.0.0
chromadb>=1.0.0
pypdf>=3.0.0
python-dotenv>=1.0.0